+++ wgpu-hal/src/gles/mod.rs
@@ -146,8 +146,6 @@ bitflags::bitflags! {
         /// Indicates that buffers used as `GL_ELEMENT_ARRAY_BUFFER` may be created / initialized / used
         /// as other targets, if not present they must not be mixed with other targets.
         const INDEX_BUFFER_ROLE_CHANGE = 1 << 5;
-        /// Indicates that the device supports disabling draw buffers
-        const CAN_DISABLE_DRAW_BUFFER = 1 << 6;
         /// Supports `glGetBufferSubData`
         const GET_BUFFER_SUB_DATA = 1 << 7;
         /// Supports `f16` color buffers
@@ -311,7 +309,6 @@ pub struct Texture {
     #[allow(unused)]
     pub format_desc: TextureFormatDesc,
     pub copy_size: CopyExtent,
-    pub is_cubemap: bool,
 }
 
 impl Texture {
@@ -332,24 +329,23 @@ impl Texture {
                 height: 0,
                 depth: 0,
             },
-            is_cubemap: false,
         }
     }
 
     /// Returns the `target`, whether the image is 3d and whether the image is a cubemap.
-    fn get_info_from_desc(desc: &TextureDescriptor) -> (u32, bool, bool) {
+    fn get_info_from_desc(desc: &TextureDescriptor) -> u32 {
         match desc.dimension {
-            wgt::TextureDimension::D1 => (glow::TEXTURE_2D, false, false),
+            wgt::TextureDimension::D1 => glow::TEXTURE_2D,
             wgt::TextureDimension::D2 => {
                 // HACK: detect a cube map; forces cube compatible textures to be cube textures
                 match (desc.is_cube_compatible(), desc.size.depth_or_array_layers) {
-                    (false, 1) => (glow::TEXTURE_2D, false, false),
-                    (false, _) => (glow::TEXTURE_2D_ARRAY, true, false),
-                    (true, 6) => (glow::TEXTURE_CUBE_MAP, false, true),
-                    (true, _) => (glow::TEXTURE_CUBE_MAP_ARRAY, true, true),
+                    (false, 1) => glow::TEXTURE_2D,
+                    (false, _) => glow::TEXTURE_2D_ARRAY,
+                    (true, 6) => glow::TEXTURE_CUBE_MAP,
+                    (true, _) => glow::TEXTURE_CUBE_MAP_ARRAY,
                 }
             }
-            wgt::TextureDimension::D3 => (glow::TEXTURE_3D, true, false),
+            wgt::TextureDimension::D3 => glow::TEXTURE_3D,
         }
     }
 }
@@ -754,7 +750,6 @@ enum Command {
         dst: glow::Texture,
         dst_target: BindTarget,
         copy: crate::TextureCopy,
-        dst_is_cubemap: bool,
     },
     CopyBufferToTexture {
         src: Buffer,