+++ wgpu-hal/src/gles/queue.rs
@@ -1,4 +1,4 @@
-use super::Command as C;
+use super::{conv::is_layered_target, Command as C};
 use arrayvec::ArrayVec;
 use glow::HasContext;
 use std::{
@@ -9,24 +9,33 @@ use std::{
 #[cfg(not(target_arch = "wasm32"))]
 const DEBUG_ID: u32 = 0;
 
-const CUBEMAP_FACES: [u32; 6] = [
-    glow::TEXTURE_CUBE_MAP_POSITIVE_X,
-    glow::TEXTURE_CUBE_MAP_NEGATIVE_X,
-    glow::TEXTURE_CUBE_MAP_POSITIVE_Y,
-    glow::TEXTURE_CUBE_MAP_NEGATIVE_Y,
-    glow::TEXTURE_CUBE_MAP_POSITIVE_Z,
-    glow::TEXTURE_CUBE_MAP_NEGATIVE_Z,
-];
-
 #[cfg(not(target_arch = "wasm32"))]
 fn extract_marker<'a>(data: &'a [u8], range: &std::ops::Range<u32>) -> &'a str {
     std::str::from_utf8(&data[range.start as usize..range.end as usize]).unwrap()
 }
 
-fn is_layered_target(target: super::BindTarget) -> bool {
+fn get_2d_target(target: u32, array_layer: u32) -> u32 {
+    const CUBEMAP_FACES: [u32; 6] = [
+        glow::TEXTURE_CUBE_MAP_POSITIVE_X,
+        glow::TEXTURE_CUBE_MAP_NEGATIVE_X,
+        glow::TEXTURE_CUBE_MAP_POSITIVE_Y,
+        glow::TEXTURE_CUBE_MAP_NEGATIVE_Y,
+        glow::TEXTURE_CUBE_MAP_POSITIVE_Z,
+        glow::TEXTURE_CUBE_MAP_NEGATIVE_Z,
+    ];
+
     match target {
-        glow::TEXTURE_2D_ARRAY | glow::TEXTURE_3D | glow::TEXTURE_CUBE_MAP_ARRAY => true,
-        _ => false,
+        glow::TEXTURE_2D => target,
+        glow::TEXTURE_CUBE_MAP => CUBEMAP_FACES[array_layer as usize],
+        _ => unreachable!(),
+    }
+}
+
+fn get_z_offset(target: u32, base: &crate::TextureCopyBase) -> u32 {
+    match target {
+        glow::TEXTURE_2D_ARRAY | glow::TEXTURE_CUBE_MAP_ARRAY => base.array_layer,
+        glow::TEXTURE_3D => base.origin.z,
+        _ => unreachable!(),
     }
 }
 
@@ -51,17 +60,12 @@ impl super::Queue {
         unsafe { gl.draw_buffers(&[glow::COLOR_ATTACHMENT0 + draw_buffer]) };
         unsafe { gl.draw_arrays(glow::TRIANGLES, 0, 3) };
 
-        let draw_buffer_count = self.draw_buffer_count.load(Ordering::Relaxed);
-        if draw_buffer_count != 0 {
+        if self.draw_buffer_count != 0 {
             // Reset the draw buffers to what they were before the clear
-            let indices = (0..draw_buffer_count as u32)
+            let indices = (0..self.draw_buffer_count as u32)
                 .map(|i| glow::COLOR_ATTACHMENT0 + i)
                 .collect::<ArrayVec<_, { crate::MAX_COLOR_ATTACHMENTS }>>();
             unsafe { gl.draw_buffers(&indices) };
-        }
-        #[cfg(not(target_arch = "wasm32"))]
-        for draw_buffer in 0..draw_buffer_count as u32 {
-            unsafe { gl.disable_draw_buffer(glow::BLEND, draw_buffer) };
         }
     }
 
@@ -127,22 +131,12 @@ impl super::Queue {
                             view.array_layers.start as i32,
                         )
                     };
-                } else if target == glow::TEXTURE_CUBE_MAP {
-                    unsafe {
-                        gl.framebuffer_texture_2d(
-                            fbo_target,
-                            attachment,
-                            CUBEMAP_FACES[view.array_layers.start as usize],
-                            Some(raw),
-                            view.mip_levels.start as i32,
-                        )
-                    };
                 } else {
                     unsafe {
                         gl.framebuffer_texture_2d(
                             fbo_target,
                             attachment,
-                            target,
+                            get_2d_target(target, view.array_layers.start),
                             Some(raw),
                             view.mip_levels.start as i32,
                         )
@@ -401,12 +395,7 @@ impl super::Queue {
                 unsafe { gl.bind_texture(dst_target, Some(dst)) };
                 let format_desc = self.shared.describe_texture_format(dst_format);
                 if is_layered_target(dst_target) {
-                    let z_offset =
-                        if let glow::TEXTURE_2D_ARRAY | glow::TEXTURE_CUBE_MAP_ARRAY = dst_target {
-                            copy.dst_base.array_layer as i32
-                        } else {
-                            copy.dst_base.origin.z as i32
-                        };
+                    let z_offset = get_z_offset(dst_target, &copy.dst_base);
 
                     match src.source {
                         wgt::ExternalImageSource::ImageBitmap(ref b) => unsafe {
@@ -415,7 +404,7 @@ impl super::Queue {
                                 copy.dst_base.mip_level as i32,
                                 copy.dst_base.origin.x as i32,
                                 copy.dst_base.origin.y as i32,
-                                z_offset,
+                                z_offset as i32,
                                 copy.size.width as i32,
                                 copy.size.height as i32,
                                 copy.size.depth as i32,
@@ -430,7 +419,7 @@ impl super::Queue {
                                 copy.dst_base.mip_level as i32,
                                 copy.dst_base.origin.x as i32,
                                 copy.dst_base.origin.y as i32,
-                                z_offset,
+                                z_offset as i32,
                                 copy.size.width as i32,
                                 copy.size.height as i32,
                                 copy.size.depth as i32,
@@ -445,7 +434,7 @@ impl super::Queue {
                                 copy.dst_base.mip_level as i32,
                                 copy.dst_base.origin.x as i32,
                                 copy.dst_base.origin.y as i32,
-                                z_offset,
+                                z_offset as i32,
                                 copy.size.width as i32,
                                 copy.size.height as i32,
                                 copy.size.depth as i32,
@@ -457,11 +446,7 @@ impl super::Queue {
                         wgt::ExternalImageSource::OffscreenCanvas(_) => unreachable!(),
                     }
                 } else {
-                    let dst_target = if let glow::TEXTURE_CUBE_MAP = dst_target {
-                        CUBEMAP_FACES[copy.dst_base.array_layer as usize]
-                    } else {
-                        dst_target
-                    };
+                    let dst_target = get_2d_target(dst_target, copy.dst_base.array_layer);
 
                     match src.source {
                         wgt::ExternalImageSource::ImageBitmap(ref b) => unsafe {
@@ -521,7 +506,6 @@ impl super::Queue {
                 src_target,
                 dst,
                 dst_target,
-                dst_is_cubemap,
                 ref copy,
             } => {
                 //TODO: handle 3D copies
@@ -550,33 +534,14 @@ impl super::Queue {
                 }
 
                 unsafe { gl.bind_texture(dst_target, Some(dst)) };
-                if dst_is_cubemap {
-                    unsafe {
-                        gl.copy_tex_sub_image_2d(
-                            CUBEMAP_FACES[copy.dst_base.array_layer as usize],
-                            copy.dst_base.mip_level as i32,
-                            copy.dst_base.origin.x as i32,
-                            copy.dst_base.origin.y as i32,
-                            copy.src_base.origin.x as i32,
-                            copy.src_base.origin.y as i32,
-                            copy.size.width as i32,
-                            copy.size.height as i32,
-                        )
-                    };
-                } else if is_layered_target(dst_target) {
+                if is_layered_target(dst_target) {
                     unsafe {
                         gl.copy_tex_sub_image_3d(
                             dst_target,
                             copy.dst_base.mip_level as i32,
                             copy.dst_base.origin.x as i32,
                             copy.dst_base.origin.y as i32,
-                            if let glow::TEXTURE_2D_ARRAY | glow::TEXTURE_CUBE_MAP_ARRAY =
-                                dst_target
-                            {
-                                copy.dst_base.array_layer as i32
-                            } else {
-                                copy.dst_base.origin.z as i32
-                            },
+                            get_z_offset(dst_target, &copy.dst_base) as i32,
                             copy.src_base.origin.x as i32,
                             copy.src_base.origin.y as i32,
                             copy.size.width as i32,
@@ -586,7 +551,7 @@ impl super::Queue {
                 } else {
                     unsafe {
                         gl.copy_tex_sub_image_2d(
-                            dst_target,
+                            get_2d_target(dst_target, copy.dst_base.array_layer),
                             copy.dst_base.mip_level as i32,
                             copy.dst_base.origin.x as i32,
                             copy.dst_base.origin.y as i32,
@@ -644,13 +609,7 @@ impl super::Queue {
                                 copy.texture_base.mip_level as i32,
                                 copy.texture_base.origin.x as i32,
                                 copy.texture_base.origin.y as i32,
-                                if let glow::TEXTURE_2D_ARRAY | glow::TEXTURE_CUBE_MAP_ARRAY =
-                                    dst_target
-                                {
-                                    copy.texture_base.array_layer as i32
-                                } else {
-                                    copy.texture_base.origin.z as i32
-                                },
+                                get_z_offset(dst_target, &copy.texture_base) as i32,
                                 copy.size.width as i32,
                                 copy.size.height as i32,
                                 copy.size.depth as i32,
@@ -662,11 +621,7 @@ impl super::Queue {
                     } else {
                         unsafe {
                             gl.tex_sub_image_2d(
-                                if let glow::TEXTURE_CUBE_MAP = dst_target {
-                                    CUBEMAP_FACES[copy.texture_base.array_layer as usize]
-                                } else {
-                                    dst_target
-                                },
+                                get_2d_target(dst_target, copy.texture_base.array_layer),
                                 copy.texture_base.mip_level as i32,
                                 copy.texture_base.origin.x as i32,
                                 copy.texture_base.origin.y as i32,
@@ -720,13 +675,7 @@ impl super::Queue {
                                 copy.texture_base.mip_level as i32,
                                 copy.texture_base.origin.x as i32,
                                 copy.texture_base.origin.y as i32,
-                                if let glow::TEXTURE_2D_ARRAY | glow::TEXTURE_CUBE_MAP_ARRAY =
-                                    dst_target
-                                {
-                                    copy.texture_base.array_layer as i32
-                                } else {
-                                    copy.texture_base.origin.z as i32
-                                },
+                                get_z_offset(dst_target, &copy.texture_base) as i32,
                                 copy.size.width as i32,
                                 copy.size.height as i32,
                                 copy.size.depth as i32,
@@ -737,11 +686,7 @@ impl super::Queue {
                     } else {
                         unsafe {
                             gl.compressed_tex_sub_image_2d(
-                                if let glow::TEXTURE_CUBE_MAP = dst_target {
-                                    CUBEMAP_FACES[copy.texture_base.array_layer as usize]
-                                } else {
-                                    dst_target
-                                },
+                                get_2d_target(dst_target, copy.texture_base.array_layer),
                                 copy.texture_base.mip_level as i32,
                                 copy.texture_base.origin.x as i32,
                                 copy.texture_base.origin.y as i32,
@@ -980,16 +925,6 @@ impl super::Queue {
                     .map(|i| glow::COLOR_ATTACHMENT0 + i)
                     .collect::<ArrayVec<_, { crate::MAX_COLOR_ATTACHMENTS }>>();
                 unsafe { gl.draw_buffers(&indices) };
-
-                if self
-                    .shared
-                    .private_caps
-                    .contains(super::PrivateCapabilities::CAN_DISABLE_DRAW_BUFFER)
-                {
-                    for draw_buffer in 0..count as u32 {
-                        unsafe { gl.disable_draw_buffer(glow::BLEND, draw_buffer) };
-                    }
-                }
             }
             C::ClearColorF {
                 draw_buffer,
@@ -1259,7 +1194,7 @@ impl super::Queue {
                         )
                     };
                     if let Some(ref blend) = *blend {
-                        unsafe { gl.enable_draw_buffer(index, glow::BLEND) };
+                        unsafe { gl.enable_draw_buffer(glow::BLEND, index) };
                         if blend.color != blend.alpha {
                             unsafe {
                                 gl.blend_equation_separate_draw_buffer(
@@ -1283,12 +1218,8 @@ impl super::Queue {
                                 gl.blend_func_draw_buffer(index, blend.color.src, blend.color.dst)
                             };
                         }
-                    } else if self
-                        .shared
-                        .private_caps
-                        .contains(super::PrivateCapabilities::CAN_DISABLE_DRAW_BUFFER)
-                    {
-                        unsafe { gl.disable_draw_buffer(index, glow::BLEND) };
+                    } else {
+                        unsafe { gl.disable_draw_buffer(glow::BLEND, index) };
                     }
                 } else {
                     unsafe {